Fix type annotation

+++ Lib/statistics.py
@@ -306,7 +306,7 @@ def _fail_neg(values, errmsg='negative value'):
         yield x
 
 
-def _isqrt_frac_rto(n: int, m: int) -> float:
+def _isqrt_frac_rto(n: int, m: int) -> int:
     """Square root of n/m, rounded to the nearest integer using round-to-odd."""
     # Reference: https://www.lri.fr/~melquion/doc/05-imacs17_1-expose.pdf
     a = math.isqrt(n // m)
